requirements.txt
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
certifi==2020.6.20
click==8.1.2
colorama==0.4.4
Deprecated==1.2.13
Flask==2.1.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.8
pytest==7.1.1
python-dotenv==0.20.0
redis==4.2.2
toml==0.10.2
tomli==2.0.1
Werkzeug==2.1.1
wincertstore==0.2
wrapt==1.14.0
